标签管理

+++ src/main/java/com/db/promote/dao/MsgTemplateMapper.java
@@ -4,17 +4,17 @@ import com.db.promote.entity.MsgTemplate;
 import org.apache.ibatis.annotations.Mapper;
 
 /**
-* Created by Mybatis Generator 2018/12/26
+* Created by Mybatis Generator 2019/01/06
 */
 @Mapper
 public interface MsgTemplateMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(MsgTemplate record);
 
     int insertSelective(MsgTemplate record);
 
-    MsgTemplate selectByPrimaryKey(Long id);
+    MsgTemplate selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(MsgTemplate record);
